+++ bin/ntx-builder/src/db/models/queries/notes.rs
@@ -77,10 +77,9 @@ pub fn insert_committed_notes(
     for note in notes {
         let row = NoteInsert {
             nullifier: conversions::nullifier_to_bytes(&note.as_note().nullifier()),
-            account_id: conversions::network_account_id_to_bytes(
-                NetworkAccountId::try_from(note.target_account_id())
-                    .expect("account ID of a network note should be a network account"),
-            ),
+            account_id: conversions::network_account_id_to_bytes(NetworkAccountId::new_trusted(
+                note.target_account_id(),
+            )),
             note_data: note.as_note().to_bytes(),
             note_id: Some(conversions::note_id_to_bytes(&note.as_note().id())),
             attempt_count: 0,