+++ crates/store/src/state/apply_block.rs
@@ -319,7 +319,7 @@ impl State {
                     block_num,
                     note_index,
                     note_id: note.id().as_word(),
-                    note_commitment: note.to_commitment(),
+                    note_commitment: note.id().as_word(),
                     metadata: *note.metadata(),
                     details,
                     attachments,